Add TypeNodeProtocol
