+++ Sources/Parser/Node/Node.swift
@@ -297,61 +297,3 @@ public class FunctionCallExpressionNode: NodeProtocol {
         self.sourceTokens = sourceTokens
     }
 }
-
-public class TypeNode: NodeProtocol {
-
-    // MARK: - Property
-
-    public let kind: NodeKind = .type
-    public var sourceTokens: [Token] { [typeToken] }
-    public let children: [any NodeProtocol] = []
-
-    public let typeToken: Token
-
-    // MARK: - Initializer
-
-    public init(typeToken: Token) {
-        self.typeToken = typeToken
-    }
-}
-
-public class PointerTypeNode: NodeProtocol {
-
-    // MARK: - Property
-
-    public let kind: NodeKind = .pointerType
-    public var sourceTokens: [Token] { referenceType.sourceTokens + [pointerToken] }
-    public var children: [any NodeProtocol] { [referenceType] }
-    public let referenceType: any NodeProtocol
-
-    public let pointerToken: Token
-
-    // MARK: - Initializer
-
-    init(referenceType: any NodeProtocol, pointerToken: Token) {
-        self.referenceType = referenceType
-        self.pointerToken = pointerToken
-    }
-}
-
-public class ArrayTypeNode: NodeProtocol {
-
-    // MARK: - Property
-
-    public var kind: NodeKind = .arrayType
-
-    public var sourceTokens: [Token] { [squareLeftToken] + elementType.sourceTokens + [squareRightToken] }
-    public var children: [any NodeProtocol] { [elementType] }
-    public let elementType: any NodeProtocol
-
-    public let squareLeftToken: Token
-    public let arraySizeToken: Token
-    public let squareRightToken: Token
-
-    public init(elementType: any NodeProtocol, squareLeftToken: Token, arraySizeToken: Token, squareRightToken: Token) {
-        self.elementType = elementType
-        self.squareLeftToken = squareLeftToken
-        self.arraySizeToken = arraySizeToken
-        self.squareRightToken = squareRightToken
-    }
-}
